+++ Introspect/Introspect.swift
@@ -186,245 +186,3 @@ private extension Array {
         return self[index]
     }
 }
-
-/// Introspection UIView that is inserted alongside the target view.
-public class IntrospectionUIView: UIView {
-    
-    required init() {
-        super.init(frame: .zero)
-        isHidden = true
-        isUserInteractionEnabled = false
-    }
-    
-    @available(*, unavailable)
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-}
-
-/// Introspection View that is injected into the UIKit hierarchy alongside the target view.
-/// After `updateUIView` is called, it calls `selector` to find the target view, then `customize` when the target view is found.
-public struct IntrospectionView<TargetViewType: UIView>: UIViewRepresentable {
-    
-    /// Method that introspects the view hierarchy to find the target view.
-    /// First argument is the introspection view itself, which is contained in a view host alongside the target view.
-    let selector: (IntrospectionUIView) -> TargetViewType?
-    
-    /// User-provided customization method for the target view.
-    let customize: (TargetViewType) -> Void
-    
-    public init(
-        selector: @escaping (UIView) -> TargetViewType?,
-        customize: @escaping (TargetViewType) -> Void
-    ) {
-        self.selector = selector
-        self.customize = customize
-    }
-    
-    public func makeUIView(context: UIViewRepresentableContext<IntrospectionView>) -> IntrospectionUIView {
-        let view = IntrospectionUIView()
-        view.accessibilityLabel = "IntrospectionUIView<\(TargetViewType.self)>"
-        return view
-    }
-
-    /// When `updateUiView` is called after creating the Introspection view, it is not yet in the UIKit hierarchy.
-    /// At this point, `introspectionView.superview.superview` is nil and we can't access the target UIKit view.
-    /// To workaround this, we wait until the runloop is done inserting the introspection view in the hierarchy, then run the selector.
-    /// Finding the target view fails silently if the selector yield no result. This happens when `updateUIView`
-    /// gets called when the introspection view gets removed from the hierarchy.
-    public func updateUIView(
-        _ uiView: IntrospectionUIView,
-        context: UIViewRepresentableContext<IntrospectionView>
-    ) {
-        DispatchQueue.main.asyncAfter(deadline: .now()) {
-            guard let targetView = self.selector(uiView) else {
-                return
-            }
-            self.customize(targetView)
-        }
-    }
-}
-
-/// Introspection UIViewController that is inserted alongside the target view controller.
-public class IntrospectionUIViewController: UIViewController {
-    required init() {
-        super.init(nibName: nil, bundle: nil)
-        view = IntrospectionUIView()
-    }
-    
-    @available(*, unavailable)
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-}
-
-/// This is the same logic as IntrospectionView but for view controllers. Please see details above.
-public struct IntrospectionViewController<TargetViewControllerType: UIViewController>: UIViewControllerRepresentable {
-    
-    let selector: (IntrospectionUIViewController) -> TargetViewControllerType?
-    let customize: (TargetViewControllerType) -> Void
-    
-    public init(
-        selector: @escaping (UIViewController) -> TargetViewControllerType?,
-        customize: @escaping (TargetViewControllerType) -> Void
-    ) {
-        self.selector = selector
-        self.customize = customize
-    }
-    
-    public func makeUIViewController(
-        context: UIViewControllerRepresentableContext<IntrospectionViewController>
-    ) -> IntrospectionUIViewController {
-        let viewController = IntrospectionUIViewController()
-        viewController.accessibilityLabel = "IntrospectionUIViewController<\(TargetViewControllerType.self)>"
-        viewController.view.accessibilityLabel = "IntrospectionUIView<\(TargetViewControllerType.self)>"
-        return viewController
-    }
-    
-    public func updateUIViewController(
-        _ uiViewController: IntrospectionUIViewController,
-        context: UIViewControllerRepresentableContext<IntrospectionViewController>
-    ) {
-        DispatchQueue.main.asyncAfter(deadline: .now()) {
-            guard let targetView = self.selector(uiViewController) else {
-                return
-            }
-            self.customize(targetView)
-        }
-    }
-}
-
-
-extension View {
-    
-    public func inject<SomeView>(_ view: SomeView) -> some View where SomeView: View {
-        return overlay(view.frame(width: 0, height: 0))
-    }
-    
-    /// Finds a `UITableView` from a `SwiftUI.List`, or `SwiftUI.List` child.
-    public func introspectTableView(customize: @escaping (UITableView) -> ()) -> some View {
-        return inject(IntrospectionView(
-            selector: { introspectionView in
-
-                // Search in ancestors
-                if let tableView = Introspect.findAncestor(ofType: UITableView.self, from: introspectionView) {
-                    return tableView
-                }
-
-                guard let viewHost = Introspect.findViewHost(from: introspectionView) else {
-                    return nil
-                }
-
-                // Search in siblings
-                return Introspect.previousSibling(containing: UITableView.self, from: viewHost)
-            },
-            customize: customize
-        ))
-    }
-    
-    /// Finds a `UIScrollView` from a `SwiftUI.ScrollView`, or `SwiftUI.ScrollView` child.
-    public func introspectScrollView(customize: @escaping (UIScrollView) -> ()) -> some View {
-        return inject(IntrospectionView(
-            selector: { introspectionView in
-                
-                // Search in ancestors
-                if let tableView = Introspect.findAncestor(ofType: UIScrollView.self, from: introspectionView) {
-                    return tableView
-                }
-                
-                guard let viewHost = Introspect.findViewHost(from: introspectionView) else {
-                    return nil
-                }
-                
-                // Search in siblings
-                return Introspect.previousSibling(containing: UIScrollView.self, from: viewHost)
-            },
-            customize: customize
-        ))
-    }
-    
-    /// Finds a `UINavigationController` from any view embedded in a `SwiftUI.NavigationView`.
-    public func introspectNavigationController(customize: @escaping (UINavigationController) -> ()) -> some View {
-        return inject(IntrospectionViewController(
-            selector: { introspectionViewController in
-                
-                // Search in ancestors
-                if let navigationController = introspectionViewController.navigationController {
-                    return navigationController
-                }
-                
-                // Search in siblings
-                return Introspect.previousSibling(containing: UINavigationController.self, from: introspectionViewController)
-            },
-            customize: customize
-        ))
-    }
-    
-    /// Finds the containing `UIViewController` of a SwiftUI view.
-    public func introspectViewController(customize: @escaping (UIViewController) -> ()) -> some View {
-        return inject(IntrospectionViewController(
-            selector: { $0.parent },
-            customize: customize
-        ))
-    }
-
-    /// Finds a `UITabBarController` from any SwiftUI view embedded in a `SwiftUI.TabView`
-    public func introspectTabBarController(customize: @escaping (UITabBarController) -> ()) -> some View {
-        return inject(IntrospectionViewController(
-            selector: { introspectionViewController in
-                
-                // Search in ancestors
-                if let navigationController = introspectionViewController.tabBarController {
-                    return navigationController
-                }
-                
-                // Search in siblings
-                return Introspect.previousSibling(ofType: UITabBarController.self, from: introspectionViewController)
-            },
-            customize: customize
-        ))
-    }
-    
-    /// Finds a `TargetView` from a `SwiftUI.View`
-    public func introspect<TargetView: UIView>(customize: @escaping (TargetView) -> ()) -> some View {
-        return inject(IntrospectionView(
-            selector: { introspectionView in
-                guard let viewHost = Introspect.findViewHost(from: introspectionView) else {
-                    return nil
-                }
-                return Introspect.previousSibling(containing: TargetView.self, from: viewHost)
-            },
-            customize: customize
-        ))
-    }
-    
-    /// Finds a `UITextField` from a `SwiftUI.TextField`
-    public func introspectTextField(customize: @escaping (UITextField) -> ()) -> some View {
-        return introspect(customize: customize)
-    }
-    
-    /// Finds a `UISwitch` from a `SwiftUI.Toggle`
-    public func introspectSwitch(customize: @escaping (UISwitch) -> ()) -> some View {
-        return introspect(customize: customize)
-    }
-    
-    /// Finds a `UISlider` from a `SwiftUI.Slider`
-    public func introspectSlider(customize: @escaping (UISlider) -> ()) -> some View {
-        return introspect(customize: customize)
-    }
-    
-    /// Finds a `UIStepper` from a `SwiftUI.Stepper`
-    public func introspectStepper(customize: @escaping (UIStepper) -> ()) -> some View {
-        return introspect(customize: customize)
-    }
-    
-    /// Finds a `UIDatePicker` from a `SwiftUI.DatePicker`
-    public func introspectDatePicker(customize: @escaping (UIDatePicker) -> ()) -> some View {
-        return introspect(customize: customize)
-    }
-    
-    /// Finds a `UISegmentedControl` from a `SwiftUI.Picker` with style `SegmentedPickerStyle`
-    public func introspectSegmentedControl(customize: @escaping (UISegmentedControl) -> ()) -> some View {
-        return introspect(customize: customize)
-    }
-}
